fixed responsivness on the navbar
