+++ frontend/src/App.jsx
@@ -3,7 +3,7 @@ import SignUpView from "./pages/Signup";
 import { Routes, Route } from "react-router-dom";
 import React from "react";
 import { Home } from "./pages/Home/HomePage";
-import { NavBar } from "./components/NavBar";
+import { NavBar } from "./components/NavBar/NavBar";
 import JobsPage from "./pages/Jobs/JobsPage";
 import Campanies from "./pages/Campanies/CampaniesPage";
 import Projects from "./pages/projects/ProjectsPage";
